renaming the main app into rest api

+++ rest_api/views.py
@@ -3,8 +3,8 @@ from rest_framework.views import APIView
 from rest_framework.parsers import JSONParser
 from django.http.response import JsonResponse
 
-from main_app.serializers import (StudentDetailsSerializer, TeacherDetailsSerializer, ParentsDetailsSerializer, DriverDetailsSerializer)
-from main_app.models import (StudentDetails, TeacherDetails, ParentsDetails, DriverDetails)
+from rest_api.serializers import (StudentDetailsSerializer, TeacherDetailsSerializer, ParentsDetailsSerializer, DriverDetailsSerializer)
+from rest_api.models import (StudentDetails, TeacherDetails, ParentsDetails, DriverDetails)
 
 
 class CreateStudent(APIView):
@@ -31,7 +31,6 @@ class SearchStudent(APIView):
 
         tutorials_serializer = StudentDetailsSerializer(tutorials, many=True)
         return JsonResponse(tutorials_serializer.data, safe=False)
-        # 'safe=False' for objects serialization
 
 
 class DeleteStudent(APIView):
@@ -39,9 +38,11 @@ class DeleteStudent(APIView):
     def delete(self, request):
 
         name = request.query_params.get('name', None)
-        delete_value = StudentDetails.objects.filter(student_name=name).delete()
-        if delete_value[0]:
-            return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
+        if name:
+            delete_value = StudentDetails.objects.filter(student_name=name).delete()
+            if delete_value[0]:
+                return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
+            return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
         return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
 
 
@@ -59,11 +60,11 @@ class CreateParent(APIView):
 class SearchParent(APIView):
 
     def get(self, request):
-        tutorials = ParentsDetails.objects.all()
 
-        title = request.query_params.get('title', None)
-        if title is not None:
-            tutorials = tutorials.filter(title__icontains=title)
+        tutorials = ParentsDetails.objects.all()
+        name = request.query_params.get('name', None)
+        if name is not None:
+            tutorials = ParentsDetails.objects.filter(student_name__icontains=name)
 
         tutorials_serializer = ParentsDetailsSerializer(tutorials, many=True)
         return JsonResponse(tutorials_serializer.data, safe=False)
@@ -71,13 +72,15 @@ class SearchParent(APIView):
 
 class DeleteParent(APIView):
 
-    def delete(self, request, pk):
+    def delete(self, request):
 
-        try:
-            tutorial = ParentsDetails.objects.get(pk=pk).delete()
-            return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
-        except ParentsDetails.DoesNotExist:
+        name = request.query_params.get('name', None)
+        if name:
+            delete_value = ParentsDetails.objects.filter(parent_name=name).delete()
+            if delete_value[0]:
+                return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
             return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
+        return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
 
 
 class CreateTeacher(APIView):
@@ -94,11 +97,11 @@ class CreateTeacher(APIView):
 class SearchTeacher(APIView):
 
     def get(self, request):
-        tutorials = TeacherDetails.objects.all()
 
-        title = request.query_params.get('title', None)
-        if title is not None:
-            tutorials = tutorials.filter(title__icontains=title)
+        tutorials = TeacherDetails.objects.all()
+        name = request.query_params.get('name', None)
+        if name is not None:
+            tutorials = TeacherDetails.objects.filter(student_name__icontains=name)
 
         tutorials_serializer = TeacherDetailsSerializer(tutorials, many=True)
         return JsonResponse(tutorials_serializer.data, safe=False)
@@ -106,20 +109,22 @@ class SearchTeacher(APIView):
 
 class DeleteTeacher(APIView):
 
-    def delete(self, request, pk):
+    def delete(self, request):
 
-        try:
-            tutorial = TeacherDetails.objects.get(pk=pk).delete()
-            return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
-        except TeacherDetails.DoesNotExist:
+        name = request.query_params.get('name', None)
+        if name:
+            delete_value = TeacherDetails.objects.filter(teacher_name=name).delete()
+            if delete_value[0]:
+                return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
             return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
+        return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
 
 
 class CreateDriver(APIView):
 
     def post(self, request, *args, **kwargs):
         tutorial_data = JSONParser().parse(request)
-        tutorial_serializer = StudentDetailsSerializer(data=tutorial_data)
+        tutorial_serializer = DriverDetailsSerializer(data=tutorial_data)
         if tutorial_serializer.is_valid():
             tutorial_serializer.save()
             return JsonResponse(tutorial_serializer.data, status=status.HTTP_201_CREATED)
@@ -129,11 +134,11 @@ class CreateDriver(APIView):
 class SearchDriver(APIView):
 
     def get(self, request):
-        tutorials = DriverDetails.objects.all()
 
-        title = request.query_params.get('title', None)
-        if title is not None:
-            tutorials = tutorials.filter(title__icontains=title)
+        tutorials = DriverDetails.objects.all()
+        name = request.query_params.get('name', None)
+        if name is not None:
+            tutorials = DriverDetails.objects.filter(student_name__icontains=name)
 
         tutorials_serializer = DriverDetailsSerializer(tutorials, many=True)
         return JsonResponse(tutorials_serializer.data, safe=False)
@@ -148,3 +153,13 @@ class DeleteDriver(APIView):
             return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
         except DriverDetails.DoesNotExist:
             return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
+
+    def delete(self, request):
+
+        name = request.query_params.get('name', None)
+        if name:
+            delete_value = DriverDetails.objects.filter(driver_name=name).delete()
+            if delete_value[0]:
+                return JsonResponse({'message': 'Tutorial was deleted successfully!'}, status=status.HTTP_204_NO_CONTENT)
+            return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
+        return JsonResponse({'message': 'The tutorial does not exist'}, status=status.HTTP_404_NOT_FOUND)
